# foodapp/static/swagger.yaml
openapi: 3.0.1
info:
  title: FastFoodFastAPI
  description: Online Food Ordering Restaurant API
  version: '0.1'
servers:
  - url: http://127.0.0.1:5000
paths:
...
  /menu/:
    get:
      description: returns all menu items
      tags:
        - Menu
      responses:
        '200':
          description: returned successfully
        '401':
          description: User unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  menu:
                    type: array
                    items:
                      type: array
                      items:
                        type: integer
              examples:
                '0':
                  value: |
                    {
                      "menu": [
                        [
                          6, 
                          "Chips and Liver", 
                          "20000", 
                          "Mon, 21 Mar 2022 11:01:04 GMT", 
                          "This dish contains cow and goat liver served with mayonaise and fries", 
                          "https://media.istockphoto.com/photos/ploughmans-lunch-with-chicken-liver-pate-picture-id488158341?b=1&k=20&m=488158341&s=170667a&w=0&h=WZzWxcPn1VWl5sqFe39fnZ1g7PYAfOMbOD3Ryjkd5No=", 
                          "10 plates"
                        ], 
                        [
                          8, 
                          "Panfried Goat", 
                          "20000", 
                          "Mon, 21 Mar 2022 11:10:44 GMT", 
                          "This dish contains cow and goat meat served with mayonaise and fries", 
                          "https://media.istockphoto.com/photos/ploughmans-lunch-with-chicken-liver-pate-picture-id488158341?b=1&k=20&m=488158341&s=170667a&w=0&h=WZzWxcPn1VWl5sqFe39fnZ1g7PYAfOMbOD3Ryjkd5No=", 
                          "10 plates"
                        ], 
                        [
                          11, 
                          "Chicken Fried Rice", 
                          "24000", 
                          "Mon, 21 Mar 2022 11:37:20 GMT", 
                          "This dish contains chicken dices with pilawo rice served with gravy", 
                          "https://images.unsplash.com/photo-1603133872878-684f208fb84b?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8Y2hpY2tlbiUyMGZyaWVkJTIwcmljZXxlbnwwfHwwfHw%3D&auto=format&fit=crop&w=500&q=60", 
                          "10 plates"
                        ], 
                        [
                          5, 
                          "Chips and Chicken", 
                          "15000", 
                          "Thu, 17 Mar 2022 15:20:46 GMT", 
                          "This dish contains marryland chicken and is served without sauce", 
                          "https://media.istockphoto.com/photos/fries-and-nuggets-picture-id458899491?b=1&k=20&m=458899491&s=170667a&w=0&h=MsJrJrn7fVmQ5_V1lL3IVyVgl2XPzjL_0HayN8bpo90=", 
                          "20 plates"
                        ], 
                        [
                          4, 
                          "Chicken Pizza", 
                          "25000", 
                          "Thu, 17 Mar 2022 15:15:57 GMT", 
                          "This dish contains mushroom soup and chicken dices ", 
                          "https://images.unsplash.com/photo-1565299624946-b28f40a0ae38?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=481&q=80", 
                          "10 boxes"
                        ]
                      ]
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
...
  /menu/{menuId}:
    get:
      description: Returns a specific menu item with its id
      tags:
        - Menu
      parameters:
        - name: menuId
          in: path
          description: ID of the menu to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returned successfully
        '401':
          description: User unauthorized
        '404':
          description: Food item not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  price:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  id:
                    type: integer
                  stock:
                    type: string
              examples:
                '0':
                  value: |
                    {
                      "description": "This dish contains mushroom soup and chicken dices ", 
                      "id": 4, 
                      "name": "Chicken Pizza", 
                      "price": "25000", 
                      "stock": "10 boxes"
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
...
  /menu/update/{id}:
    post:
      description: Updates a specific menu item
      parameters:
      - in: formData
        name: food_name
        description: Food name
        type: string
      - in: formData
        name: food_price
        description: Food price
        type: int64
      - in: formData
        name: food_description
        description: Food description
        type: string 
      - in: formData
        name: food_stock_quantity
        description: Food stock quantity
        type: string   
      - in: formData
        name: image_url
        description: Food image url
        type: string   
      tags:
      - Menu
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                food_name:
                  type: string
                food_price:
                  type: integer
                image_url:
                  type: string
                food_description:
                  type: string
                food_stock_quantity:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "food_name":"Chips and Liver",
                      "food_price":20000,
                      "food_description":"This dish contains cow and goat liver served with mayonaise and fries",
                      "food_stock_quantity":"12 plates",
                      "image_url":"https://media.istockphoto.com/photos/ploughmans-lunch-with-chicken-liver-pate-picture-id488158341?b=1&k=20&m=488158341&s=170667a&w=0&h=WZzWxcPn1VWl5sqFe39fnZ1g7PYAfOMbOD3Ryjkd5No="
                  }
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  food_name:
                    type: string
                  food_price:
                    type: integer
                  image_url:
                    type: string
                  food_description:
                    type: string
                  message:
                    type: string
                  food_stock_quantity:
                    type: string
              examples:
                '0':
                  value: |
                    {
                      "food_description": "This dish contains cow and goat liver served with mayonaise and fries", 
                      "food_name": "Chips and Liver", 
                      "food_price": 20000, 
                      "food_stock_quantity": "12 plates", 
                      "image_url": "https://media.istockphoto.com/photos/ploughmans-lunch-with-chicken-liver-pate-picture-id488158341?b=1&k=20&m=488158341&s=170667a&w=0&h=WZzWxcPn1VWl5sqFe39fnZ1g7PYAfOMbOD3Ryjkd5No=", 
                      "message": "Food Updated"
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000    
  /menu/create:
    post:
      description: Creates a new menu item
      parameters:
      - in: formData
        name: food_name
        description: Food name
        type: string
      - in: formData
        name: food_price
        description: Food price
        type: integer
      - in: formData
        name: food_description
        description: Food description
        type: string 
      - in: formData
        name: food_stock_quantity
        description: Food stock quantity
        type: string   
      - in: formData
        name: image_url
        description: Food image url
        type: string  
      tags:
        - Menu
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                food_name:
                  type: string
                food_price:
                  type: integer
                image_url:
                  type: string
                food_description:
                  type: string
                food_stock_quantity:
                  type: string
            examples:
              '0':
                value: |2-
                   {
                             "food_name": "Fish Fillet",
                              "food_description":"This dish contains fish fillet and is served with any other food dish of your choice like rice, irish potatoes, mashed potatoes and vegetable salads",
                             "food_price": 30000,
                            "image_url":  "https://media.istockphoto.com/photos/grilled-halibut-with-spinach-leeks-and-pine-nuts-picture-id503337620?b=1&k=20&m=503337620&s=170667a&w=0&h=L-LJlhhsNNQVrY65U6Ogj4vRZyC8QitUJshHl-mgZ_I=",
                             "food_stock_quantity": "15 plates"
                   }
      responses:
        '500':
          description: Food added successfully
        '401':
          description: User unauthorized
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
...
components:
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://yourauthurl.com
          scopes:
            scope_name: Enter your scopes here